support interntrain framework for ascend npu
